Refactor binfmt functions from utility.cpp
