+++ src/felix86/common/binfmt.hpp
@@ -1,0 +1,4 @@
+#pragma once
+#include <string>
+
+bool unregister_binfmt_misc(const std::string& path);
