+++ src/felix86/common/utility.hpp
@@ -157,9 +157,6 @@ const std::string& felix86_cpuinfo();
 
 bool check_if_privileged_executable(const std::filesystem::path& path);
 
-// TODO: move me to new binfmt.hpp file along with binfmt_misc function
-bool unregister_binfmt_misc(const std::string& path);
-
 std::string disassemble_one(u64 address);
 
 void felix86_crash_and_burn();
